Represent locations in sig ASTs with keys
Summary:
When the abstract locations flag is turned on, we now change the sig ASTs so that they represent their locations using keys, with the concrete location values stored in a separate table in the shared heap.

As of right now, we don't safely convert these back to concrete locations anywhere, so turning this flag on will cause Flow to immediately crash. However, with a few upcoming changes, abstract locations will work as expected for toy examples, as evidenced by D15426761. More changes will be needed before this is production-ready.

Reviewed By: avikchaudhuri

Differential Revision: D14091203

fbshipit-source-id: 0a2f25ae7f64a20fe4c2eba8e13211b13a167c0c