[trusted-types] Enable trust inference on all type annotations
Summary:
This diff enables trust inference by making all type annotations (and targets of casts) trust-inferred. Most of the action is in the `convert` function of `type_annotation`, where all types are now given trust inference variables that are initialized to unknown.

Several new test cases are included, and a lot of old test cases now get different results. This is because prior to this change, a type annotation like `var x: number` would treat `number` as untrusted, and if it flowed into a trusted type--`var y: $Trusted<number>`--an error would be raised. Now, this is totally fine unless `number` has had an explicitly tainted type or `any` flow into it.

Reviewed By: dsainati1

Differential Revision: D14669361

fbshipit-source-id: 0d221e1d9856e755bde4a171ddc9042ecb7422d8