use build_revision instead of build_id_ohai for build id
Summary:
`flow version` includes a `build_id` field which contains the commit hash and commit time at which the binary was built. for example:

```
303a72d565fb10134a60018aa485f9a031f034be Mon Jun 10 17:24:39 2019
```

the time is unnecessary; the hash is unique (modulo SHA-1 attacks... famous last words!) and a hash is easier to use than a string with spaces and a hard-to-parse localized date in it.

this isn't used for the client/server handshake, which already uses `build_revision`.

Reviewed By: gabelevi

Differential Revision: D15760949

fbshipit-source-id: aa31dde804ddb192de5a921f8f2879269beda374