[hack] Queue up and forward file change notifications
Summary: The serverless IDE needs to be informed of changes to files on disk so that it can update its in-memory data structures, such as the naming table. We use the LSP `workspace/didChangeWatchedFiles` notification to subscribe to file change events. We just need to forward them to the actual process which does IDE stuff.

Reviewed By: pittsw

Differential Revision: D15529983

fbshipit-source-id: 5da5d534ace4dd27490a66d0bceab90e4d4ef680