Optimize collect_of_types using FullyResolved node
Summary:
This visitor walks tvars, trying to find any unresolved nodes. Since a
FullyResolved node will never contain anything unresolved, we can avoid walking
those bounds entirely.

This is an important optimization, as the structure of exports can be quite
complex, seeing as structure from imported modules often appears in the exports
of the importing one.

I believe we can use this information to further remove the need to store a
type graph in sig contexts -- and store all tvars from dep cxs in the finished
set of the active cx's type graph -- which adds unnecessary overhead to full
type resolution. However, due to MergedT it is possible for full type
resolutino to span across merge jobs, which doesn't really make any sense and
is going to take a lot of deep through to tease apart.

Reviewed By: avikchaudhuri

Differential Revision: D10528155

fbshipit-source-id: 6d5d8aae995edeae83d21a1ac40daff19f205e1d