Prevent choice kit uses from leaking into dependent merge steps
Summary:
As noted in the comment inline, this should never be necessary. However, the
current system puts non-0->1 tvars into the type graph, which can prevent
the speculative constraints from firing.

If this leaves residue in signature contexts (via MergedT), then we might try
to explore parts of the type graph from a previous merge step. Currently we
handle this by copying over the previous merge step, but this is unnecessarily
wasteful.

Follow-up diffs will remove this wasteful step and further optimize the type
graph, so that D10226436 has a chance of landing. Once that change lands, then
it will become feasible to pin down type parameters such that they are actually
0->1 always.

Reviewed By: avikchaudhuri

Differential Revision: D10385423

fbshipit-source-id: c3f785116762df4d4ebfddd29e12b23f66d4b295