Track root ids during concretization
Summary:
The id of any arbitrary OpenT might correspond to a Goto node in the tvar
graph. Since a Goto node is, by definition, equivalent to it's corresponding
Root node, we should use the Root node ids to track our seen set.

In other words, if we've already seen the Root node corresponding to a given
Goto node, we don't need to further concretize -- there is nothing new to
learn.

Reviewed By: avikchaudhuri

Differential Revision: D10432420

fbshipit-source-id: 14d5c06efe8d6c6994ece9d2d073a1dd0398baae