Use Hashtbl.add instead of Hashtbl.replace in subst cache
Summary:
These are chaining hash tables, so each hash key corresponds to a linked list of
nodes. The replace function walks the linked list to find a matching node and
replace it, or add a new one if no new node matches.

We always check for membership first, so this traversal is unnecessary.

Reviewed By: panagosg7

Differential Revision: D15815647

fbshipit-source-id: f8488586d167779d525fc1ad331a98a9e65b107e