Drop reason from collect_of_type
Summary:
We threaded this around everywhere to satisfy a single conditional, which
checks if the reason is instantiable.

It's problematic to have behavior that depends on this reason, because it makes
it difficult to share results between two calls to collect_of_types. I want to
replace the "seen" accumulator with the state of the type graph instead.

Reviewed By: avikchaudhuri

Differential Revision: D10439856

fbshipit-source-id: 0b0166a851522594b82c269071d67c476de3dbee