Track root ids during speculation
Summary:
Two type variables are equivalent if they have the same root id. Before
speculative match, we walk types and collect open+unresolved type variables,
adding their ids to the speculation case.

If we see an action that involves that tvar, we will defer its evaluation, to
avoid side-effects until a single branch can be selected. Unless we track the
root ids, we will fail to defer evaluation if we see a constraint involving an
equivalent tvar with a different id.

Reviewed By: avikchaudhuri

Differential Revision: D10173071

fbshipit-source-id: e87b97bee334fc22dc591cd3af712f3b84d0b830