fix semver comparison of ranges
Summary:
given a range like `^0.101.0`, which decomposes into `>=0.101.0 <0.102.0`, the version `0.102.0-rc` should *not* match, even though it is technically less than 0.102.0.

`0.102.0-rc` should only match the range if at least one of the comparators in that range has matching major minor and patch, and a prerelease tag.

as another example, `0.103.0-beta` does match the range `>= 0.101.0 <0.103.0-rc`, because `0.103.0-beta` and `0.103.0-rc` have identical major/minor/version, and `beta` < `rc`. but `0.102.0-beta` does _not_ match.

this is all super confusing and not part of the semver spec, but matches what node-semver does.

Reviewed By: dsainati1

Differential Revision: D15799757

fbshipit-source-id: ce1bb3077f81064d9c723859a3682749abecc34b