[hack] Server docblock command should work even without position information
Summary:
LSP autocomplete includes two requests, `textDocument/completion` and `completionItem/resolve`.

The first request is intended to be fast and quickly give you a list of symbols you can choose from.  The second request is intended to give you details about a single symbol when you select it in the autocomplete UI.

Previously, our responses to `textDocument/completion` contained position information, which took 100-200ms to calculate.  This made autocomplete very slow, but `completionItem/resolve` required that the position information be sent in the API request in order to do its work.  If you failed to send it, nothing would display.

This diff enables the `completionItem/resolve` API to work correctly even without position information, by simply looking up the symbol.  As a result, autocomplete can be fast and resolve can continue to display docblocks.

This was solved by:
* Add new methods inside SymbolIndex that locate a symbol in the repository
* Adding a new server RPC call, `LOCATE_SYMBOL`, which allows a symbol to be looked up if location is not known.  This is needed because Client can't do the work on its own; it has to contact server to do the work.
* Splitting `DOCBLOCK_AT` into two parts: locating the symbol and then secondarily fetching the docblock.  If the JSON data contains a valid position, we use that position; but if the JSON request does not, then we use `LOCATE_SYMBOL`.

Limitations:
* If you sent bad data in the "data" part of the resolve request, you will still get bad data back from the resolve response.

Reviewed By: pittsw

Differential Revision: D15781687

fbshipit-source-id: 0cc43f1cc1953b8e27f4cce8cee5fe489d2af6cc