Drop unnecessary pseudo_init_declared_type api
Summary:
Flow allows the following code:

```
var x: string;
x = "foo";
```

Bindings can be in one of three states: Undeclared, Declared, and Initialized.
This API was used to bring the binding `x` into the Declared state. In all
cases, the binding will already be in the Declared state, so this call is
unnecessary.

If the binding is a `var`, the `bind_var` call during the decl pass will create
a scope entry in the Declared state.

For `let` and `const` bindings, the decl pass creates a scope entry in the
Undeclared state. The `declare_var` call, which is a noop for var bindings, will
perform the Undeclared -> Declared transition.

Reviewed By: panagosg7

Differential Revision: D15345052

fbshipit-source-id: e5d3b24cefa7c2aff09ddaad6a419594a888d8fb