Rework variable declaration and initialization code
Summary:
Before this diff, the decl pass did a surprising amount of work processing
variable declarations. Notably, it would actually call into the destructuring
helper and initialize destructured bindings.

This change moves all that work into the expr pass where it belongs. This
simplifies the decl pass which now only needs to create tvars, which the
inference pass will constrain.

For annotated destructuring bindings, the tvar created in the decl pass will be
unified with an AnnotT type which constraints subsequent writes. Note that there
is still a bug here involving destructured union-like types, which will be
resolved in a follow-up.

Reviewed By: panagosg7

Differential Revision: D15421902

fbshipit-source-id: b3acb61018e5519a511fb70d88271d8c114e6600