flow-remove-types: Support comment types in input
Summary:
Fixes an issue where comment type annotations of the form `/*: T */` would not be removed entirely - specifically, the closing `*/` would remain in the output, leading to parse errors.

v1 of `flow-remove-types` used to preserve Flow comment types, probably because the version of Babylon we used knew nothing about them. v2, which uses the Flow parser, ran into a problem with the current Flow AST: The `TypeAnnotation` node starts at the beginning of the comment, but ends before the end of the underlying `MultiLineComment`. In this diff, we use the start position to detect whether a `TypeAnnotation` originates in a comment, and if so skip processing it. This is safe because downstream parsers will just see a comment there.

Reviewed By: mroch

Differential Revision: D15897583

fbshipit-source-id: 302df28256490d38db26ad904c63aba5a4a98596