[autofix][insert-type] add ability to update file in place
Summary: This commit adds the `--in-place` flag to the autofix command. This flag writes to the `--path` argument by default and if not specified writes to the input file. If no input file or path argument is given and the `--in-place` flag is given then an error is reported. If the flag is not given the command result is output to `stdout`.

Reviewed By: mvitousek

Differential Revision: D15881852

fbshipit-source-id: df78857eeb5615882faf7b09ace252b234c6a8b2