Evaluate for-in/for-of head expression in the correct scope
Summary:
Before this change, Flow evaluated the head expression of a for-in/for-of
statement in parent scope. This is rarely an observable difference, but consider
a case like this:

```
const x: Array<number> = [];
for (const x of x) {
  x;
}
```

In the above example, Flow would erroneously consider the right-hand side
expression to be an identifier that resolves to `Array<number>` from the parent
scope. The binding created in the left declaration and used in the body was
erroneously ascribed the type `number`.

Per the spec [0] (see it's uses in for-in and for-of [1]), the right-hand
side expression should be evaluated in the same scope as the left declaration.
This means that the identifier expression actually resolves to the binding
declared in the left declaration, which is an error because the binding x has
not yet been declared.

I also noticed that both of these code paths were doing unnecessary work. They
were evaluating the right-hand side expression but also calling the refinement
function `predicates_of_condition`, which also evaluates the expression. For
now, I fixed this by only using `predicates_of_condition`.

Note that the spec [0] describes a simpler logic that does not map exactly onto
the refinement logic currently in use.

0. https://www.ecma-international.org/ecma-262/7.0/#sec-runtime-semantics-forin-div-ofheadevaluation-tdznames-expr-iterationkind
1. https://www.ecma-international.org/ecma-262/7.0/#sec-for-in-and-for-of-statements

Reviewed By: dsainati1

Differential Revision: D15924022

fbshipit-source-id: 719555d7366652c817d6a20b99eacdfae2829622