[def-inst-prop-assign] copy ssa_builder
Summary: I am going to generalize the state of the ssa_builder so that we can use it to determine when a property is definitively assigned in the constructor. For now, I just copied the file verbatim. Once, I have made my changes, I will integrate my work with the existing ssa_builder. Working in a separate module will make it easier to get the code working for property assignment without having to deal with code that depends on the actual ssa_builder.

Reviewed By: jbrown215

Differential Revision: D15941581

fbshipit-source-id: ef629024352d1f6796d7da057801eed409fc63ee