[types-first] Fix suppressions in lazy mode
Summary:
In types-first, errors and suppressions from the merge phase were cleared
out earlier. For errors we only need to keep the ones raised during the
checking phase. This doesn't hold for suppressions. We need to keep all
of them around instead of using a stripped down version.

Before, suppressions were populated correctly for streaming errors, but not
for non-streaming ones.

Reviewed By: avikchaudhuri

Differential Revision: D15917590

fbshipit-source-id: 8462f33fb3bbacc58a9d7b31527f5e1d9999c2dd