fix types-first perf regression over classic
Summary:
We already knew that in types-first, when two classes are defined in two files that are in a dependency cycle, we might hit artificial incompatibilities between two "versions" of the same class, one that is computed in the combined signature of these files during merge, and the other that is reached through the signature when checking each of these files separately. This was avoided by using locations as nominal ids for classes, so that there was no essential difference between the two versions that would cause nominal subtyping to fail.

Unfortunately, there is a subtle variation of the same problem with structural subtyping: not behavioral, but in terms of performance. When two versions of the same structural type are compared in types-first, some common fast paths might not be hit. Thus, our optimizations need to be more robust in recognizing equalities (by looking past benign differences).

This diff adds one such optimization.

Note for the future: we might be able to avoid accidental regressions from classic to types-first by not processing the same types twice, once in merge and once in check: given a file to check that is part of a dependency cycle, we can "diff" the AST and the sig AST of the file and only check the diff.

Reviewed By: panagosg7

Differential Revision: D15955212

fbshipit-source-id: c8b9bb2cec8befcffaafc12a0416617770ee4b99