Correctly sighash abstract locations
Summary: This is one of the rare circumstances where it is desirable to reach into the representation of an `ALoc.t` to determine whether it has an underlying concrete or abstract representation. By sighashing based on the key of an abstract location, rather than its underlying concrete value, we can make good on the promise of the abstract locations project to avoid rechecks when only locations have changed.

Reviewed By: avikchaudhuri

Differential Revision: D14091208

fbshipit-source-id: dc9efc3c660508b87a0872affc2178b993110f7a