Use lookup tables to convert from ALoc.t to Loc.t during error collation
Summary:
This is the other side of D14091203, which changes sig ASTs to use `ALoc.t`s represented by keys, rather than concrete locations.

When we get locations out of the other side of the typechecker, we have to convert them back to `Loc.t` safely. Currently we use `ALoc.to_loc_exn` everywhere, which raises if the `ALoc.t` is not just a `Loc.t` under the hood. For abstract locations to work, we need to use lookup tables to convert them back when they are represented by keys.

Reviewed By: gabelevi

Differential Revision: D14091206

fbshipit-source-id: e344f7c6b7023d6a43ba179401391013ffbd500d