fixing a badly positioned error related to MapTypeT
Summary:
Background
Without this diff, the newly added test program blames the definition at line 1, instead of the use at line 3.
The cause is that the use_op of `MapTypeT` is not properly recognized in the `use_op_of_use_t` function.
Now it's fixed.

Without this diff, when reporting an error with a `use_op`, the upper bound is discarded.
Normally discarding the upper bound is not a problem, because for most cases the `use_op` contains the same information as the upper bound does.
But in the newly added test, not including the upper bound results in a less informative error,
since the upper bound here is not included in the `use_op`.
This could cause more problems as we are improving error messages.
So, in this diff, the upper bound is added to the error generation.

Reviewed By: samwgoldman

Differential Revision: D15943220

fbshipit-source-id: 8df397e4cbe804249ec911c363b261dee1735045