account for diff in hh_fixmes to decide whether to retypecheck dependency
Summary:
Changes in the HH_FIXMEs attached to some class elements would fail to have the dependencies retypechecked.
As in:
```
// file1.php
final class :myxhp {
  attribute
    /* HH_FIXME[4110] Type mismatch */
    int myattr;
}
```
```
// file2.php
function test(): void {
  $x = <myxhp myattr="something"/>;
}
```
where removing the HH_FIXME would fail to show an error in file2.php.

To fix that, this diff add an attribute `elt_fixme_codes` in the `element` type in Decl_defs.

Reviewed By: 2BitSalute

Differential Revision: D15899012

fbshipit-source-id: 693796c52abb0919ecca04e34e17b607a2340f15