fixing a badly positioned error related to PolyT
Summary:
Background:
For the newly added test, Flow used to blame the definitions at line 18 and line 30. Those are badly positioned errors.

Cause:
In `__flow`, when the lower bound is `PolyT`, the `use_op` of the upper bound is not always properly propagated,
because the fall-through line of pattern matching introduces a `unknown_use`.
For the newly added test, the upper bound is `ReactKitT` (a specialized case of `ReactKitT` is handled, but it doesn't cover other cases) and its `use_op` is lost.

Change:
Added the `use_op` of the upper bound to the fall-through line.

Reviewed By: panagosg7

Differential Revision: D15970624

fbshipit-source-id: 620b5fed9504fee13bd1365941dad73f9255a4db