Drop now meaningless "strict" distinction from merge code
Summary:
A long, long time ago, there wasn't very much modulatity in Flow. To combat
this, a "strict" version of merge as written in parallel to the existing
version. The strict version enforced that the input positions of exports were
annotated.

The "non-strict" version of this code is long gone. We rely on the modularity
in a very fundamental way. It is no longer necessary to call these things strict
in the absence of any alternate mode.

It's also somewhat confusing, since the term "strict" has become very
overloaded.

Reviewed By: dsainati1

Differential Revision: D16040509

fbshipit-source-id: 44bf19d82cbcee7f1db4bdaeb5ad854f628ef7ac