Include abstract location tables for the current component in the context
Summary:
## Why this is needed

There are many places where we have optimizations or even observable typechecking behavior depend on location comparisons. Unfortunately, we may end up with a mix of abstract and concrete locations in these comparisons for various reasons:

* When we eagerly filter lints during typechecking, we need to check which line the lint location is on. This can happen during merge (when we are dealing with only abstract locations), so we need to be able to look up their underlying value. See D15724740, D15748449.
* We use locations as unique identifiers for classes. In cycles, we can end up with both abstract and concrete locations which refer to the same class. It would be incorrect to allow these to compare as unequal. See D15725195.
* We perform typechecking optimizations based on locations. See D15955212 for an example (reasons contain locations, so this call to polymorphic compare relies on the underlying representation of the location).

Because we need access to the concrete value of locations for the current component being typechecked, we need to make the abstract location tables available for all files in the current component.

## Why this is safe

This makes concrete locations for the current component available while typechecking that component. The benefit of abstract locations is that their concrete values can change without re-typechecking. Therefore, the risk here is that the concrete locations that we have access to in the typechecker may change in a way that affects typechecking, without actually causing the typechecker to re-run on these files. I believe this cannot happen:

Under types-first (which is required for abstract locations to be enabled), cycles only exist during merge. During check, each file is checked individually, against the type signatures for its dependencies that were computed during merge.

First, I will justify this for the check phase. In the check phase, all components have size 1. Therefore, only the abstract location tables for the current file getting typechecked are available. In order for the locations in a file to change, the file itself must change. Therefore, it is impossible for the locations to change without also re-running the typechecking that relied upon them.

Next, I will justify this for the merge phase. Here, cycles exist, so components may have any size.

Cycles, in the merge phase, are always typechecked together as a unit.

Recheck opts are applied only **after** a component is merged, and are used to optimize which downstream components are merged, and to optimize the check phase. They cannot cause only part of a component to be merged (Sam Goldman, I'm not sure how this may change with your proposed work on recheck opts).

Therefore, if any file in a cycle changes (including adding or removing dependency edges), we will remerge the entire cycle. Because of that, there is no way for the concrete locations of the files in a component to change without the typechecker getting re-run on all of those files.

Reviewed By: samwgoldman

Differential Revision: D15724643

fbshipit-source-id: 337d9177497483be505707a7f61841df2f5dc0da