[hack] Support sorting keys for determinic multiline JSONs
Summary:
The pretty output of `Hh_json` module doesn't *really* deliver on promise
to be deterministic; i.e., it lacks a well-defined order of fields in objects.
This blocks migration of tests to a different library, more specifically,
it's **impossible** to swap out JSON-printing OCaml code for a Rust code
(without some wild assumption on insertion-order and the like),
which is heavily used in FactsParser and related tests.

The expected output of tests will be updated in a follow-up diff.

Differential Revision: D16045465

fbshipit-source-id: 0880b6575aa31fdf107172c74900946214f650c6