better calculation of all dependent files
Summary:
Previously, the set of all dependent files was computed by considering code dependencies throughout. This is sub-optimal with types-first, since a dependent of `f` is not necessarily one whose code recursively depends on `f`, but one whose code depends on a file whose signature recursively depends on `f`'s signature. This diff makes that optimization.

Moreover, previously the calculation of direct dependents and all dependents was paired, and done before the dependency graph was updated. But whereas the set of direct dependents is special in that various kinds of changes that require rechecking of direct dependents is not necessarily captured in the updated dependency graph, the set of all dependents essentially follows from the direct dependents and the updated dependency graph. This diff makes the split in that calculation, and removes some redundancy related to closure computation along the way.

Reviewed By: samwgoldman

Differential Revision: D15669877

fbshipit-source-id: f25fa74bd9a5d7bd17744ec5034424e78bd91a97