Use ALoc table for the current component when eagerly lint filtering in Flow_js
Summary:
The main explanation for this approach is available in D15724643.

This specific case is due to the dynamic export lint. It gets fired in the context optimizer during merge, and then we hit this codepath when we try to filter it.

Because lint filtering involves checking if a location is within a range, we need the concrete location to be available.

Reviewed By: dsainati1

Differential Revision: D15724740

fbshipit-source-id: cf4ac1cca6c47872ac57d6d8988846c69ed39509