ALoc.to_loc_with_tables utility
Summary:
This utility makes it easy to convert an `ALoc.t` to a `Loc.t` when you have a map of tables. It picks out the correct table for you. This is convenient because we have just such a map in the context.

Requested by Daniel Sainati to avoid duplication in upcoming diffs.

Reviewed By: dsainati1

Differential Revision: D16060422

fbshipit-source-id: d307b9964f3d6eed19a4584b576f7a3e8189af38