Object.freeze creates derivable types
Summary: Derivable tvars enable `assert_ground` to recurse into those tvars with positive polarity. This is useful for type-level operations where the output is essentially derived from the input: if the input is sufficiently annotated, the output need not be. In particular, calling `Object.freeze` on an object creates an object with such a derived type.

Reviewed By: samwgoldman

Differential Revision: D16065493

fbshipit-source-id: 73901793f5bfbf5df8a9602ba1cccb2c901fa52d