Use `ALoc.concretize_equal` on locations used as IDs in the typechecker
Summary: This is required wherever one `ALoc.t` for a single file may be abstract (represented by a key), and another may be concrete, otherwise our options are to either crash or incorrectly determine that the locations are not equal. Specifically, this can happen in cycles, as exercised by the test in this diff. See D15724643 for more context.

Reviewed By: avikchaudhuri

Differential Revision: D15725195

fbshipit-source-id: b904b7dd2508a49821f9277d85a63aeacf78ed8c