`compare` and `equal` variants for ALoc.t to safely compare keyed locs with concrete locs
Summary:
In certain cases, we can end up wanting to compare a concrete location to an abstract location (both of type `ALoc.t`, with different underlying representations) in the same file. Naively, they would compare as not equal, but that's incorrect. In order to do a meaningful comparison in this case, we must convert the abstract one to be concrete. See D15725195 for one such example.

This diff adds compare functions that take a map of available tables to use to compare an abstract location with a concrete one. Of course, if the needed table is not present in the map, these functions will raise a runtime error.

Reviewed By: avikchaudhuri

Differential Revision: D15725179

fbshipit-source-id: d529f20afb383aaaa09b8e148d8ec58b43ea956b