Add `--types` flag to `flow cycle`
Summary:
Type dependencies are different than value ones. For example the import of `foo` in
```
// file b.js
import {foo} from 'a';
foo();
module.exports = class A {};
```
does not affect the type signature of file a.js since it does not affect its exports' type.

The edge
```
a.js --> b.js
```
is only a value dependency but not a type one.

This affects the behavior of `flow cycle` in that a cycle computed based on type
dependencies only might differ from one that accounts for both.

In Flow classic mode we took all dependencies into account to compute `flow cycle`.
This behavior had regressed in types-first since `flow cycle` only accounted for
type dependencies. This change restores the behavior for the default case
of this command, and also adds an option `--types` to only account for type dependencies
when computing cycles.

In addressing performance issues due to cycles in types-first, one should try to break
type-dependencies instead of the combination of the two.

Reviewed By: gabelevi

Differential Revision: D16018435

fbshipit-source-id: c059bc480be536e7185c8e92f8bc52cc3944c8b4