Make get-def work with abstract locations
Summary:
We need to be able to safely resolve abstract locations to concrete ones, so we need access to the lookup tables, which are stored in shared memory.

I'm a little bit concerned about the performance of possible repeated deserializations of the tables from shared memory, but a similarly-naive approach worked fine for errors. I believe there is a cache built into the shared memory module, which probably helps. I'll optimize this in the future if needed.

Reviewed By: gabelevi

Differential Revision: D16042683

fbshipit-source-id: 395b41ad6919912d3cd262be0287d7d04e89adb4