`loc_of_aloc` utility function
Summary:
I found myself repeating this snippet while converting IDE services to be safe with respect to abstract locations. It's convenient to have it as a function because it allows `loc_of_aloc ~reader` to be a drop-in replacement for `ALoc.to_loc_exn`.

I'm putting it in this folder so as not to introduce additional dependencies to a more generic utilities module.

Reviewed By: panagosg7

Differential Revision: D16053028

fbshipit-source-id: c14e2efbe840f2f6d2ff6e1470cb4dc2d5e2f941