Use ppx deriving eq to generate an equality function for `Reason.t`
Summary:
For various reasons, we would like to avoid using polymorphic compare. It's particularly urgent in this case, however, because otherwise identical reasons may contain locations which are logically equal but have different representations. This causes them to compare as not equal, frustrating important optimizations.

Specifically, we may be comparing locations where one is represented by a key whereas another has a concrete representation. We need to be able to look up the concrete value of such locations in order to properly do the comparison. See also D15725179 and D15725195.

Luckily, the `virtual_reason` type is polymorphic on locations already, so Deriving requires us to pass in a location equality function to its generated function. This allows us to easily use `ALoc.concretize_equal` rather than defaulting to `ALoc.compare`.

Reviewed By: mroch

Differential Revision: D16018234

fbshipit-source-id: bfda3c014882ba55cb280f7ecbab8f65eeae926e