fixing a badly positioned error related to PolyT
Summary:
Background:
Without this diff, the newly added test generates an error that blames the definition at line 6.
That is a badly positioned error.

Fix:
In `__flow`, when the lower bound is `PolyT` and the upper takes the form `ConcretizeTypeAppsT(op1, (_, op2, _), (_, op3, _), _)`, if `op1` is not unknown, then use `op1` instead of `op2` or `op3`, because `op1` is the location of the use.

Reviewed By: panagosg7

Differential Revision: D16069393

fbshipit-source-id: ba0da5ea486a584b426f30bc9568c16d51ae36a1