[autofix][insert-type] improve error messages
Summary: This is an incremental improvement and documentation of the current error messages for the autofix insert-type tool. Note the error numbers is UNKOWN_ERROR for now and most error messages are not that great of a user experience. These are both things that I would like to improve in the near future.

Reviewed By: panagosg7

Differential Revision: D16112047

fbshipit-source-id: 310df1332fbc92b1f959c86bba353fb5eb1a5a4f