[def-inst-prop-assign] bind properties for ssa_builder
Summary:
We need to feed the properties to `ssa_builder` as `Hoister.Bindings`. Since `ssa_builder` only track reads and writes to identifiers, we are going to treat `this.someProperty` as an identifier within the `ssa_builder`. (Doing it this way also makes sense semantically, as `this` should not really exist until the constructor exits, so we can think of `this.someProperty` as just a fancy way of referring to the identifier `someProperty`.) We also want to prefix the identifier with `this.` to avoid name clashes, e.g. with let bindings in the constructor.

We also need to catch abrupt completion exceptions from `ssa_builder` to prevent Flow from crashing. `ssa_builder` deals with abrupt completions (i.e. `return`, `throw`, `break`, and `continue`) by raising an exception. Typically, this is not an issue because specific statements are configured to expect these exceptions (so that they can merge ssa environments). Since we are feeding `ssa_builder` the body of the constructor as if it were a full program, it is not expecting `return` and `throw` to appear (except for maybe in a function definition nested inside the constructor). Therefore, we should catch all `return`s and `throw`s that bubble up to our entry point into `ssa_builder`. In the next Dif in the stack, I am going to make sure the ssa environments get merged correctly.

Note that this Diff runs the subclass of `ssa_builder`, `property_assignment`, on the constructor body but does not use it for emitting errors yet (that comes later). For the scope of this Diff, we just want to prepare the inputs to the `property_assignment` class and make sure that we can run it on the constructor without causing Flow to crash.

Reviewed By: jbrown215

Differential Revision: D16046578

fbshipit-source-id: b0902675b90d41a5232248c6377a44b02b48cccd