fix top_cycle metric
Summary:
I'm assuming the top_cycle metric is supposed to be the largest cycle by number of members. it's currently the one whose leader is last alphabetically (`f2 > f1`).

it's also unnecessarily wastefully computed. we don't need to fold over a map to create a list of tuples and then map, filter and fold over that; we can just compute the top cycle directly.

Reviewed By: samwgoldman

Differential Revision: D16187349

fbshipit-source-id: 9e855e8a5411777cb9b4d217d8faae678be53ec2