[def-inst-prop-assign] error on reads to uninited properties
Summary: We want to emit an error if a property is read from before it is initialized, even if it is eventually initialized later on in the constructor. `ssa_builder` contains a data structure called `values` that stores the preceding write (or list of potential writes in the case of branching) for every read of a "tracked binding." By iterating over `values`, we can check that for every read from a property, the list of writes that could "reach" that read does not contain `uninitialized`. If it contains `uninitialized`, there must be some code path where the property is not initialized at the point when it is read, and so we emit an error.

Reviewed By: jbrown215

Differential Revision: D16050094

fbshipit-source-id: e41ddb431fafc1009922a21f61450af9fdce0884