[def-inst-prop-assign] fix eval order of assignments
Summary: Assignments have different evaluation orders depending on whether the LHS is an identifier or expression. As mentioned in D16046578, properties are semantically more similar to identifiers than expressions. Therefore, assignments involving properties need to be evaluated in the same order that an identifier would be evaluated. The one difference in the evaluation order of properties in comparison to identifiers is that we do not consider update assignments (like `this.someProperty += 3`) as writes to the property. I made this choice to improve the clarity of error messages.

Reviewed By: jbrown215

Differential Revision: D16049974

fbshipit-source-id: aa0fbef88abc32bf69fcb6c1b209ad2854d095d7