[def-inst-prop-assign] let this escape ctor when all properties inited
Summary: It is actually okay for `this` to escape the constructor when all properties are definitively initialized. In order to check that condition, we must integrate the `this_in_constructor_finder` code with the `property_assignment` class. We can mostly copy and paste the code between the classes. One subtle difference is that once the `this_in_constructor` code found a method call in the constructor it did not bother checking the arguments to the method while we should be doing that. While doing it the old way was technically sound (since it only masked errors when it was already emitting an error), it affected the error diff strangely. Namely, had the this Diff kept the old behavior, it would have reduced the number of `ReadFromUninitializedProperty`s when it should have only reduced the number of `ThisBeforeEverythingInitialized`s and `MethodCallBeforeEverythingInitialized`s.

Reviewed By: jbrown215

Differential Revision: D16067286

fbshipit-source-id: 5c3c412e2855816b1f846c2fa966d91ea1eb27d8