[def-inst-prop-assign] merge ssa env for return and throw
Summary:
We need special handling for when `return` and `throw` appear in the constructor. Instead of just catching the `AbruptCompletion.exn` that is raised, we want to merge the environments from the abrupt completions with the ssa environment from the "normal" completion.

It is very important that `expecting_return_or_throw` is nested inside `with_bindings` and not the other way around. When `with_bindings` exits, it pops the ssa environment which would destroy the environment from the "normal" completion.

Reviewed By: jbrown215

Differential Revision: D16088807

fbshipit-source-id: 20be26b82818299d2ae4c1929dc74c01288bcfdb