[def-inst-prop-assign] error when this escapes constructor
Summary: Emit an error when `this` escapes the constructor. I eventually want to integrate the `this_in_constructor_finder` class with the `property_assignment` class. This will allow us to make the analysis more precise by allowing `this` to escape once all properties are initialized rather than never allowing it to escape the constructor. This Diff establishes a good initial baseline of soundness for the planned improvements.

Reviewed By: jbrown215

Differential Revision: D16053264

fbshipit-source-id: 5118a9dd577a9513cbc4174181702fadba00c866